+++ packages/tsurlfilter/src/rules/declarative-converter/errors/converter-options-errors/index.ts
@@ -1,14 +1,14 @@
 import { EmptyOrNegativeNumberOfRulesError } from './empty-or-negative-number-of-rules-error';
-import { NegativeNumberOfRegexpRulesError } from './negative-number-of-regexp-rules-error';
+import { NegativeNumberOfRulesError } from './negative-number-of-rules-error';
 import { ResourcesPathError } from './resources-path-error';
 
 type ConverterOptionsError = EmptyOrNegativeNumberOfRulesError
-| NegativeNumberOfRegexpRulesError
+| NegativeNumberOfRulesError
 | ResourcesPathError;
 
 export {
     type ConverterOptionsError,
     EmptyOrNegativeNumberOfRulesError,
-    NegativeNumberOfRegexpRulesError,
+    NegativeNumberOfRulesError,
     ResourcesPathError,
 };