+++ packages/tsurlfilter/src/rules/declarative-converter/errors/converter-options-errors/negative-number-of-rules-error.ts
@@ -1,7 +1,7 @@
 /**
  * Describes error when maximum number of rules is less than 0.
  */
-export class NegativeNumberOfRegexpRulesError extends Error {
+export class NegativeNumberOfRulesError extends Error {
     /**
      * Describes error when maximum number of rules is less than 0.
      *
@@ -10,9 +10,9 @@ export class NegativeNumberOfRegexpRulesError extends Error {
     constructor(message: string) {
         super(message);
 
-        this.name = 'NegativeNumberOfRegexpRulesError';
+        this.name = this.constructor.name;
 
         // For proper work of the "instanceof" operator
-        Object.setPrototypeOf(this, NegativeNumberOfRegexpRulesError.prototype);
+        Object.setPrototypeOf(this, NegativeNumberOfRulesError.prototype);
     }
 }