+++ packages/tsurlfilter/src/rules/declarative-converter/rule-set.ts
@@ -42,6 +42,13 @@ export interface IRuleSet {
     getRulesCount(): number;
 
     /**
+     * Number of converted declarative unsafe rules.
+     *
+     * @returns Number of converted declarative unsafe rules.
+     */
+    getUnsafeRulesCount(): number;
+
+    /**
      * Number of converted declarative regexp rules.
      *
      * @returns Number of converted declarative regexp rules.
@@ -129,6 +136,7 @@ type SerializedRuleSetLazyData = zod.infer<typeof serializedRuleSetLazyDataValid
 
 const serializedRuleSetDataValidator = zod.strictObject({
     regexpRulesCount: zod.number(),
+    unsafeRulesCount: zod.number().optional(),
     rulesCount: zod.number(),
     ruleSetHashMapRaw: zod.string(),
     badFilterRulesRaw: zod.string().array(),
@@ -187,10 +195,16 @@ export class RuleSet implements IRuleSet {
 
     /**
      * Number of converted declarative rules.
+     *
      * This is needed for the lazy version of the rule set,
      * when content not loaded.
      */
     private readonly rulesCount: number = 0;
+
+    /**
+     * Converted declarative unsafe rules.
+     */
+    private readonly unsafeRulesCount: number = 0;
 
     /**
      * Converted declarative regexp rules.
@@ -242,6 +256,7 @@ export class RuleSet implements IRuleSet {
      *
      * @param id Id of rule set.
      * @param rulesCount Number of rules.
+     * @param unsafeRulesCount Number of unsafe rules.
      * @param regexpRulesCount Number of regexp rules.
      * @param ruleSetContentProvider Rule set content provider.
      * @param badFilterRules List of rules with $badfilter modifier.
@@ -250,6 +265,7 @@ export class RuleSet implements IRuleSet {
     constructor(
         id: string,
         rulesCount: number,
+        unsafeRulesCount: number,
         regexpRulesCount: number,
         ruleSetContentProvider: RuleSetContentProvider,
         badFilterRules: IndexedNetworkRuleWithHash[],
@@ -257,6 +273,7 @@ export class RuleSet implements IRuleSet {
     ) {
         this.id = id;
         this.rulesCount = rulesCount;
+        this.unsafeRulesCount = unsafeRulesCount;
         this.regexpRulesCount = regexpRulesCount;
         this.ruleSetContentProvider = ruleSetContentProvider;
         this.badFilterRules = badFilterRules;
@@ -266,6 +283,11 @@ export class RuleSet implements IRuleSet {
     /** @inheritdoc */
     public getRulesCount(): number {
         return this.rulesCount || this.declarativeRules.length;
+    }
+
+    /** @inheritdoc */
+    public getUnsafeRulesCount(): number {
+        return this.unsafeRulesCount;
     }
 
     /** @inheritdoc */
@@ -547,6 +569,7 @@ export class RuleSet implements IRuleSet {
 
         const data: SerializedRuleSetData = {
             regexpRulesCount: this.regexpRulesCount,
+            unsafeRulesCount: this.unsafeRulesCount,
             rulesCount: this.rulesCount,
             ruleSetHashMapRaw: this.rulesHashMap.serialize(),
             // TODO: Remove .getText() completely